Implement ULX hooks
We implemented ULX hooks to be able to catch ban and unban events by ULX and use those in MBSync, this way when someone gets banned using ULX we can ban him cross network too. One issue is the way ULX stores banData, because we are unable to define if the ban should be on all servers or not using this method. Another issue is that we need to use a regex to be able to get the steamid of the banning admin ( come on ULX team, are you serious? Why do you save the Nickname + Steamid in the same variable )